fix(wireguard): add iptables rule to allow traffic to 10.43.0.0/16
This commit adds an `iptables` rule to the `PostUp` section of the WireGuard configuration. This rule allows traffic destined for the `10.43.0.0/16` network to be forwarded out through the default interface. This is necessary to ensure that internal network traffic is correctly routed when the WireGuard tunnel is active.

Additionally, a corresponding `iptables` rule is added to the `PreDown` section to remove the rule when the tunnel is disconnected, preventing potential network conflicts.